requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.1.1
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.3
datasets==2.15.0
dill==0.3.7
diskcache==5.6.3
dnspython==2.4.2
emoji==2.8.0
entrypoints==0.4
evaluate==0.4.1
exceptiongroup==1.2.0
extra-streamlit-components==0.1.60
Faker==20.1.0
favicon==0.7.0
filelock==3.13.1
filetype==1.2.0
fonttools==4.46.0
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.1
htbuilder==0.6.2
huggingface-hub==0.19.4
idna==3.6
importlib-metadata==6.11.0
InstructorEmbedding==1.0.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
langchain==0.0.348
langchain-core==0.0.12
langdetect==1.0.9
langsmith==0.0.69
llama_cpp_python==0.2.20
loguru==0.7.2
lxml==4.9.3
Markdown==3.5.1
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
mdurl==0.1.2
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
mysql-connector-python==8.2.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
packaging==23.2
pandas==2.1.3
passlib==1.7.4
Pillow==10.1.0
pinecone-client==2.2.4
protobuf==4.21.12
pyarrow==14.0.1
pyarrow-hotfix==0.6
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pymdown-extensions==10.5
pyparsing==3.1.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-iso639==2023.6.15
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.5.2
referencing==0.31.1
regex==2023.10.3
requests==2.31.0
responses==0.18.0
rich==13.7.0
rpds-py==0.13.2
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
st-annotated-text==4.0.1
streamlit==1.29.0
streamlit-authenticator==0.2.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-embedcode==0.1.2
streamlit-extras==0.3.5
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.2
streamlit-lottie==0.0.5
streamlit-option-menu==0.3.6
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
torch==2.1.1
torchvision==0.16.1
tornado==6.4
tqdm==4.66.1
transformers==4.35.2
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
unstructured==0.11.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
win32-setctime==1.1.0
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.3
zipp==3.17.0
